# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API для управления астрономическими исследованиями планет",
        "title": "Astronomy Research API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/research/research-cart": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Возвращает информацию о текущем черновике исследования пользователя",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "researches"
                ],
                "summary": "Получить корзину исследования",
                "responses": {
                    "200": {
                        "description": "Данные корзины исследования",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Неверный запрос",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/research/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Возвращает полную информацию об исследовании включая планеты",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "researches"
                ],
                "summary": "Получить исследование по ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID исследования",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Данные исследования с планетами",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Неверный ID",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "Доступ запрещен",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Исследование не найдено",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/research/{id}/change-research": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Обновляет данные исследования",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "researches"
                ],
                "summary": "Изменить исследование",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID исследования",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Новые данные исследования",
                        "name": "research",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/apitypes.ResearchJSON"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Обновленное исследование",
                        "schema": {
                            "$ref": "#/definitions/apitypes.ResearchJSON"
                        }
                    },
                    "400": {
                        "description": "Неверные данные",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Исследование не найдено",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/research/{id}/delete-research": {
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Выполняет логическое удаление исследования",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "researches"
                ],
                "summary": "Удалить исследование",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID исследования",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Статус удаления",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Неверный запрос",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "Доступ запрещен",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Исследование не найдено",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/research/{id}/finish": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Изменяет статус исследования (только для модераторов)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "researches"
                ],
                "summary": "Модерировать исследование",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID исследования",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Новый статус",
                        "name": "status",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/apitypes.StatusJSON"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Результат модерации",
                        "schema": {
                            "$ref": "#/definitions/apitypes.ResearchJSON"
                        }
                    },
                    "400": {
                        "description": "Неверный запрос",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "Доступ запрещен",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Исследование не найдено",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/research/{id}/form": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Переводит исследование в статус \"formed\"",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "researches"
                ],
                "summary": "Сформировать исследование",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID исследования",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Сформированное исследование",
                        "schema": {
                            "$ref": "#/definitions/apitypes.ResearchJSON"
                        }
                    },
                    "400": {
                        "description": "Неверный запрос",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "Доступ запрещен",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Исследование не найдено",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/researches": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Возвращает исследования с возможностью фильтрации по датам и статусу",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "researches"
                ],
                "summary": "Получить список исследований",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Начальная дата (YYYY-MM-DD)",
                        "name": "from-date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Конечная дата (YYYY-MM-DD)",
                        "name": "to-date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Статус исследования",
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список исследований",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/apitypes.ResearchJSON"
                            }
                        }
                    },
                    "400": {
                        "description": "Неверный формат даты",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "apitypes.ResearchJSON": {
            "type": "object",
            "properties": {
                "creator_login": {
                    "type": "string"
                },
                "date_create": {
                    "type": "string"
                },
                "date_finish": {
                    "type": "string"
                },
                "date_form": {
                    "type": "string"
                },
                "date_research": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "moderator_login": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "apitypes.StatusJSON": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}